backend_python/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
python-multipart
passlib[bcrypt]
email-validator
bcrypt
